Improve cpp tests - null cases and more floating point edge cases.

+++ src/main/cpp/src/hash.cuh
@@ -25,7 +25,7 @@
 
 namespace spark_rapids_jni {
 
-constexpr int SPARK_DEFAULT_XXHASH64_SEED = 42;
+constexpr int64_t SPARK_DEFAULT_XXHASH64_SEED = 42;
 
 /**
  * @brief Converts a cudf decimal128 value to a java bigdecimal value.
